+++ templates/access/logged-in-only.php
@@ -8,7 +8,8 @@ if (!defined('ABSPATH')) {
 }
 ?>
 <div class="loopis-message information">
-	<p>🚧 Du behöver vara medlem för att se något här.</p>
-	<p><span class="big-link">💳 <a href="<?php echo esc_url(home_url('/shop/?option=membership-stripe')); ?>">Betala medlemskap</a></span></p>
+	<p>🚧 Du behöver vara inloggad för att se något här.</p>
 	<?php include LOOPIS_THEME_DIR . '/templates/links/go-back.php'; ?>
 </div>
+
+<?php include LOOPIS_THEME_DIR . '/templates/links/log-in-button.php'; ?>