+++ templates/links/log-in.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * Show button to log in.
+ */
+ 
+if (!defined('ABSPATH')) {
+    exit; // Exit if accessed directly
+}
+?>
+
+<p><span class="big-link"><a href="<?php echo esc_url(home_url('/log-in')); ?>">👤 Logga in</a></span></p>